% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starsim.R
\name{init_starsim}
\alias{init_starsim}
\title{Initialize Starsim}
\usage{
init_starsim(..., envname = "r-reticulate", required = FALSE)
}
\value{
null
}
\description{
Perform the steps needed to initialize Starsim: install Miniconda, create a virtual
environment ("r-reticulate"), and install Starsim into it
}
\examples{
\dontrun{
init_starsim()
}
}
