% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starsim.R
\name{init_starsim}
\alias{init_starsim}
\title{Initialize Starsim}
\usage{
init_starsim(..., envname = "r-starsim", required = FALSE)
}
\description{
Perform the steps needed to initialize Starsim: install Miniconda, create a virtual
environment ("r-starsim"), and install Starsim into it
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
init_starsim()
\dontshow{\}) # examplesIf}
}
