% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starsim.R
\name{load_starsim}
\alias{load_starsim}
\title{Load Starsim}
\usage{
load_starsim(envname = "r-starsim", required = FALSE)
}
\description{
Load all components of the Starsim environment, including: \code{ss} (all
Starsim functionality); NumPy (\code{np}), pandas (\code{pd}), Sciris (\code{sc}), and
Matplotlib (\code{plt}), for additional Python functionality; and some of
the core Starsim classes (e.g. \code{Sim}, \code{Disease}). After installation,
this is the only function that needs to be called from R-Starsim;
everything else is accessed via \code{ss}.
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
load_starsim()
sim <- ss$Sim(diseases='sis', networks='random')
sim$run()
\dontshow{\}) # examplesIf}
}
