% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starsim.R
\name{reinstall_starsim}
\alias{reinstall_starsim}
\title{Reinstall Starsim}
\usage{
reinstall_starsim(..., envname = "r-reticulate")
}
\value{
null
}
\description{
Reinstalls Starsim into the current environment ("r-reticulate" by default)
}
\examples{
\dontrun{
reinstall_starsim()
}
}
