% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starsim.R
\name{reinstall_starsim}
\alias{reinstall_starsim}
\title{Reinstall Starsim}
\usage{
reinstall_starsim(..., envname = "r-reticulate")
}
\description{
Reinstalls Starsim into the current environment ("r-reticulate" by default)
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
reinstall_starsim()
\dontshow{\}) # examplesIf}
}
